tests/submission/PhilipTaylor/canvas/size.attributes.reflect.setidlzero.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.reflect.setidlzero</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.<a href="index.size.attributes.reflect.html">reflect</a>.setidlzero</h1>
<p class="desc">Setting IDL attributes to 0 updates IDL and content attributes</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.reflect">size.reflect</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Setting IDL attributes to 0 updates IDL and content attributes");
_addTest(function(canvas, ctx) {

canvas.width = 0;
canvas.height = 0;
_assertSame(canvas.getAttribute('width'), '0', "canvas.getAttribute('width')", "'0'");
_assertSame(canvas.getAttribute('height'), '0', "canvas.getAttribute('height')", "'0'");
_assertSame(canvas.width, 0, "canvas.width", "0");
_assertSame(canvas.height, 0, "canvas.height", "0");


});
</script>

<div id="log"></div>
</body></html>
